#feat: add payment method

+++ src/Usage.kt
@@ -2,10 +2,10 @@ import java.util.*
 import kotlin.concurrent.thread
 
 class Usage {
-    fun buyTickets() {
-        try {
-            for (personNumber in 1..numberOfPeople) {
-                thread(start = true) {
+    fun buyTickets() = try {
+        for (personNumber in 1..numberOfPeople) {
+            thread(start = true) {
+                try {
                     val database = Database(getConnection()!!)
                     val random = Random()
                     val events = database.getEvents()
@@ -13,16 +13,22 @@ class Usage {
                     val areas = database.getAreas(selectedEvent.id)
                     val selectedArea = areas[random.nextInt(areas.size)]
                     val ticket = database.getTicket(selectedArea.id)
-                    val person = database.getPerson("Person $personNumber")
 
-                    database.updateTicket(person!!,ticket!!)
+                    if(ticket != null) {
+                        val person = database.getPeople("Person $personNumber")[0]
 
-                    database.destroy()
+                        database.updateTicket(person,ticket!!)
+
+                        database.destroy()
+                    }
+
+                } catch (e : Exception) {
+                    System.err.println(e.javaClass.name + ": " + e.message)
                 }
             }
-        } catch (e: Exception) {
-            System.err.println(e.javaClass.name + ": " + e.message)
-            System.exit(0)
         }
+    } catch (e: Exception) {
+        System.err.println(e.javaClass.name + ": " + e.message)
+        System.exit(0)
     }
 }
